+++ NEWS.md
@@ -1,3 +1,5 @@
+# zipangu (development version)
+
 # zipangu 0.2.3
 
 ## Fix