+++ NEWS.md
@@ -6,9 +6,13 @@
     - Determine if the string is a prefecture string by `is_prefecture()`. 
 - `str_jnormalize`: New function for preprocessing Japanese characters following the rules of 'neologd' (@paithiov909, #36).
 
+## Fix
+
+- Fix the behaviors of `jholiday_spec()` when any NAs, NULL, and empty strings are supplied (@paithiov909, #38).
+
 ## Enhancement
 
-- `jholiday_spec()` is now memoised internally. This change makes its second and later calls faster. Also, now `name` is acceptable with multiple values. (@paithiov909, #38).
+- `jholiday_spec()` is now memoised internally. This change makes its second and later calls faster. Also, multiple values of `name` are now acceptable. (@paithiov909, #38).
 
 # zipangu 0.2.3